Pulls in:
- credential: Metadata/Description→Record/Metadata rename (51baa36f)
- workspace: delete nebula-config, replace CLI loader with figment
- workspace: replace custom InterfaceVersion with semver::Version
- schema: [lib] doctest = false from cargo shear (rejected — we keep
  the 9 doctests introduced in Task 24)

Conflict resolution: crates/schema/Cargo.toml — kept our bench_resolve
and bench_lookup sections, dropped main's [lib] doctest = false since
nebula-schema now ships with working doctests.

+++ crates/sdk/src/action.rs
@@ -34,7 +34,7 @@ pub struct ActionBuilder {
     key: ActionKey,
     name: String,
     description: String,
-    version: (u32, u32),
+    version: (u64, u64),
 }
 
 impl ActionBuilder {
@@ -59,8 +59,8 @@ impl ActionBuilder {
         self
     }
 
-    /// Set the interface version.
-    pub fn with_version(mut self, major: u32, minor: u32) -> Self {
+    /// Set the interface version (major, minor components; patch defaults to 0).
+    pub fn with_version(mut self, major: u64, minor: u64) -> Self {
         self.version = (major, minor);
         self
     }
